requirements.txt
# Python dependencies for Rin Computer Control System

# Core automation
pywin32
pyautogui>=0.9.54
Pillow>=10.0.0
mss>=9.0.0

# Configuration
pyyaml>=6.0.1

# HTTP client for llama-server API
requests>=2.31.0

# Backend server
fastapi>=0.109.0
uvicorn>=0.27.0

# Testing
pytest>=7.4.0

# Voice interaction
openwakeword>=0.5.0
useful-moonshine-onnx>=0.1.0  # Moonshine ASR - 2x better accuracy than Vosk
sounddevice>=0.4.6

# Discord integration (optional)
discord.py>=2.3.0

# WebSocket real-time updates
python-socketio>=5.10.0

# Encryption for secrets management
cryptography>=42.0.0
